tests: Generate test steps depending on OPTLEVEL
The generated `test.xl.mk` files are where the option given in `$OPTLEVEL` are
recorded. When they are extraced from, say, a `CMD=` entry in the test, the
output is wrong if we do it in a way that is independent from `OPTLEVEL`.

This was routinely causing the `option-bool-variant` test to fail because the
initial value for the `-O` option was incorrect.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>