bytecode: Enter leaf constants without a type
When we enter a leaf tree, we want the classifciation to happen correctly. The
machine type in that case could get in the way, i.e. we could have a natural
`RunValue` and a `Natural *` pointer, and that would result in a `RunValue`
having the value of the pointer in its `as_natural` field.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>