bytecode: Set `INVALID_OPCODE` as opcode 0
Use `INVALID_OPCODE` to take slot of opcode 0, which is otherwise used by
`tree_cast`. This resulted in logic errors for code that was relying on having a
non-zero opcode as an indication of having a valid opcode. One manifestation was
a crash in the `tests/01.Evaluation/05-kind.xl`, with messages such as:

```
[48470 0.456400] opcode_error: Nonexistent builtin 0x7f8cb7c95ae8 ["tree_cast"]
[71068 0.493923] opcode_error: Nonexistent builtin 0x7f8cb7c95ae8 ["tree_cast"]
[71937 0.495949] opcode_error: Nonexistent builtin 0x7f8cb7c95ae8 ["tree_cast"]
[72822 0.498377] opcode_error: Nonexistent builtin 0x7f8cb7c95ae8 ["tree_cast"]
```

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>