tests: Accelerate fibonacci test failure
The Fibonacci test is an example of a "slow recursive", so the failure is
extremely slow to arrive if we run it with a high value. Since the goal of the
test is to detect failure, not to do it quickly, it makes sense to run the test
with a much lower value.

That being said, running the test with a high value without having to wait for a
long time would be highly desirable.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>