Fix internal error in 01.Evaluation/variable-result.xl
In the following case:

    A : integer := 17
    function Index(I:text) as variable integer is A
    write Index("A")

we used to get an error like:

    Index(A)=v.xl:37:17: Expected a text value, got 17

This was broken by commit:55d44b54846b7198a81eb5d5c67bce758f6beade where we
added the inherited cached values to the current evaluation.

The reason is that in `Bindings::Do(Infix *)`, we were returning the result [A :
integer := 17] (lifted from the evaluation cache), but since this is part of a
cast, we were then blindingly accepting the evaluation value, 17, as input to a
text argument for the [write T:text] form. This was caught later because this is
a native form.

By instead returning the original cast, we let the caller know that we found a
cast that needs to be evaluated. This means that the [text] type is correctly
ruled out. However, the output is still not good, since we now return a [tree]
that can be written direclty by the [write T] form. As a result, the test that
raised the issue, `tests/01.Evaluation/variable-result.xl`, shows the following
incorrect result:

    A=17 B=2 C=1969
    Index(A)=A:integer := 17
    Index(B)=B:integer := 2
    Index(C)=C:integer := 1969

instead of

    A=17 B=2 C=1969
    Index(A)=17
    Index(B)=2
    Index(C)=1969

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>